Rename package from jpy-tools to splurge-tools, update version to 0.1.12, and adjust project metadata. Update README and all relevant imports in tests to reflect the new package name. Remove deprecated modules and ensure all references are consistent with the new structure.
