--- jpy_tools/py.typed
@@ -1,2 +1,0 @@
-# This file marks the jpy_tools package as type-annotated
-# See PEP 561 for more information
