+++ splurge_tools/py.typed
@@ -1,0 +1,2 @@
+# This file marks the splurge_tools package as type-annotated
+# See PEP 561 for more information
